alignment with navit git done

+++ src/rebuild_all.sh
@@ -3,10 +3,14 @@
 debug="OFF"
 franca="OFF"
 html="OFF"
+clean=0
 
-while getopts dfh opt
+while getopts cdfh opt
 do
 	case $opt in
+	c)
+		clean=1
+		;;
 	d)
 		debug="ON"
 		;;
@@ -18,7 +22,8 @@ do
 		;;
 	\?)
 		echo "Usage:"
-		echo "$0 [-dfh]"
+		echo "$0 [-cdfh]"
+		echo "-c: Rebuild with clean"
 		echo "-d: Enable the debug messages"
 		echo "-f: Build using the Franca interfaces"
 		echo "-h: Enable migration to the html based hmi"
@@ -27,8 +32,16 @@ do
 done
 set -e
 
-echo 'clean up the build folder'
-find ./build ! -name '*.cbp' -type f -exec rm -f {} +
+if [ "$clean" = 1 ]
+then
+	echo 'clean up the build folder and regenerate all the stuff'
+	if [ -d "./build" ]
+	then
+		find ./build ! -name '*.cbp' -type f -exec rm -f {} +
+	fi
+else
+	echo 'just build without generation of the hmi'
+fi
 
 mkdir -p build
 cd build
@@ -36,26 +49,37 @@ mkdir -p navigation
 cd navigation
 mkdir -p navit
 cd navit
-mkdir -p navit
-cd navit
 
 echo 'build navit'
-cmake -DDISABLE_QT=1 -DSAMPLE_MAP=0 -Dvehicle/null=1 -Dgraphics/qt_qpainter=0 ../../../../navigation/src/navigation/navit/navit/
+if [ "$clean" = 1 ]
+then
+	cmake -DDISABLE_QT=1 -DSAMPLE_MAP=0 -Dvehicle/null=1 -Dgraphics/qt_qpainter=0 ../../../navigation/src/navigation/navit/
+fi
 make
-cd ../../
+cd ../
 
 echo 'build navigation'
-cmake -DWITH_DEBUG=$debug ../../navigation/src/navigation
+if [ "$clean" = 1 ]
+then
+	cmake -DWITH_DEBUG=$debug ../../navigation/src/navigation
+fi
 make
 cd ..
 
 echo 'build fsa'
-cmake -DWITH_HTML_MIGRATION=$html -DWITH_FRANCA_DBUS_INTERFACE=$franca -DCOMMONAPI_DBUS_TOOL_DIR=$COMMONAPI_DBUS_TOOL_DIR -DCOMMONAPI_TOOL_DIR=$COMMONAPI_TOOL_DIR -DWITH_DEBUG=$debug ../
+
+if [ "$clean" = 1 ]
+then
+	cmake -DWITH_HTML_MIGRATION=$html -DWITH_FRANCA_DBUS_INTERFACE=$franca -DCOMMONAPI_DBUS_TOOL_DIR=$COMMONAPI_DBUS_TOOL_DIR -DCOMMONAPI_TOOL_DIR=$COMMONAPI_TOOL_DIR -DWITH_DEBUG=$debug ../
+fi
 make
 cd ../
 
-echo 'generate the hmi for gdp theme'
-cd script
-./prepare.sh -i ../hmi/qml/Core/gimp/gdp-theme/800x480
-cd ../
+if [ "$clean" = 1 ]
+then
+	echo 'generate the hmi for gdp theme'
+	cd script
+	./prepare.sh -i ../hmi/qml/Core/gimp/gdp-theme/800x480
+	cd ../
+fi
